requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==9.8.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
